fix issues with album art in background not changing properly
Also resolve issues with the background flashing black while the image loads.

In certain cases the album art in the player won't have updated by the time we process a song change. We now add a Mutation Observer for the album art to detect when it changes after we've sampled it.